+++ src/tbox/prefix/assert.h
@@ -237,7 +237,7 @@ __tb_extern_c_enter__
 #endif
 
 /// assert: noimpl
-#define tb_assert_noimpl()                                  tb_assertf(0, "noimpl")
+#define tb_assert_noimpl()                                      tb_assertf(0, "noimpl")
 
 /*! the static assert
  *
@@ -247,7 +247,13 @@ __tb_extern_c_enter__
  *
  * @endcode
  */
-#define tb_assert_static(x)                                 do { typedef int __tb_static_assert__[(x)? 1 : -1]; __tb_volatile__ __tb_static_assert__ __a; tb_used_ptr((tb_cpointer_t)(tb_size_t)__a); } while(0)
+#if __tb_has_feature__(c_static_assert)
+#   define tb_assert_static(x)      _Static_assert(x, "")
+#elif defined(TB_COMPILER_IS_GCC) && TB_COMPILER_VERSION_BE(4, 6)
+#   define tb_assert_static(x)      _Static_assert(x, "")
+#else
+#   define tb_assert_static(x)      do { typedef int __tb_static_assert__[(x)? 1 : -1]; __tb_volatile__ __tb_static_assert__ __a; tb_used_ptr((tb_cpointer_t)(tb_size_t)__a); } while(0)
+#endif
 
 /* //////////////////////////////////////////////////////////////////////////////////////
  * declaration